refactor(emit): split oversized namespace and source tests
