+++ crates/tsz-emitter/src/transforms/namespace_es5_ir.rs
@@ -50,8 +50,12 @@
 //! })(A || (A = {}));
 //! ```
 
+#[path = "namespace_es5_ir_const_enum.rs"]
+mod namespace_es5_ir_const_enum;
 #[path = "namespace_es5_ir_helpers.rs"]
 mod namespace_es5_ir_helpers;
+#[path = "namespace_es5_ir_source.rs"]
+mod namespace_es5_ir_source;
 use namespace_es5_ir_helpers::*;
 
 use std::cell::{Cell, RefCell};
@@ -301,260 +305,6 @@ impl<'a> NamespaceES5Transformer<'a> {
             return std::collections::HashSet::new();
         };
         collect_namespace_function_scope_reference_names(self.arena, innermost_body)
-    }
-
-    /// Extract leading comments from source text that fall within [`from_pos`, `to_pos`) range.
-    /// Returns `IRNode::Raw` nodes since the text already includes comment delimiters.
-    fn extract_comments_in_range(&self, from_pos: u32, to_pos: u32) -> Vec<IRNode> {
-        let source_text = match self.source_text {
-            Some(t) => t,
-            None => return Vec::new(),
-        };
-        let mut result = Vec::new();
-        for c in &self.comment_ranges {
-            if c.pos >= from_pos && c.end <= to_pos {
-                let text = c.get_text(source_text);
-                if !text.is_empty() {
-                    result.push(IRNode::Raw(text.to_string().into()));
-                }
-            }
-            if c.pos >= to_pos {
-                break; // Comments are sorted by position
-            }
-        }
-        result
-    }
-
-    /// Skip whitespace and comments forward from `pos` to find the actual token start.
-    /// Returns the position of the first non-trivia character.
-    fn skip_trivia_forward(&self, pos: u32, end: u32) -> u32 {
-        super::emit_utils::skip_trivia_forward(self.source_text, pos, end)
-    }
-
-    /// Find the position after the code content of an erased statement (interface/type alias).
-    /// Scans forward with brace-depth tracking to find the closing `}` or `;`.
-    /// This is needed because `node.end` includes trailing trivia that may contain
-    /// comments belonging to the next statement.
-    fn find_code_end_of_erased_stmt(&self, node_pos: u32, node_end: u32) -> u32 {
-        let source_text = match self.source_text {
-            Some(t) => t,
-            None => return node_end,
-        };
-        let bytes = source_text.as_bytes();
-        let end = (node_end as usize).min(bytes.len());
-        let mut i = node_pos as usize;
-        let mut brace_depth: i32 = 0;
-        let mut found_brace = false;
-
-        while i < end {
-            // Skip over comment ranges
-            let pos = i as u32;
-            let mut skipped_comment = false;
-            for c in &self.comment_ranges {
-                if c.pos <= pos && pos < c.end {
-                    i = c.end as usize;
-                    skipped_comment = true;
-                    break;
-                }
-                if c.pos > pos {
-                    break; // comments sorted by position
-                }
-            }
-            if skipped_comment {
-                continue;
-            }
-
-            match bytes[i] {
-                b'{' => {
-                    brace_depth += 1;
-                    found_brace = true;
-                }
-                b'}' => {
-                    brace_depth -= 1;
-                    if found_brace && brace_depth == 0 {
-                        return (i + 1) as u32;
-                    }
-                }
-                b';' if brace_depth == 0 && !found_brace => {
-                    // Type alias without braces: type Foo = number;
-                    return (i + 1) as u32;
-                }
-                b'\'' | b'"' => {
-                    // Skip string literal
-                    let quote = bytes[i];
-                    i += 1;
-                    while i < end && bytes[i] != quote {
-                        if bytes[i] == b'\\' {
-                            i += 1;
-                        }
-                        i += 1;
-                    }
-                    if i < end {
-                        i += 1;
-                    }
-                    continue;
-                }
-                _ => {}
-            }
-            i += 1;
-        }
-
-        node_end
-    }
-
-    /// Extract standalone comments (on their own line) within [`from_pos`, `to_pos`).
-    /// Unlike `extract_comments_in_range`, this filters out trailing comments
-    /// that share a line with code — only comments on their own line are returned.
-    fn extract_standalone_comments_in_range(&self, from_pos: u32, to_pos: u32) -> Vec<IRNode> {
-        let source_text = match self.source_text {
-            Some(t) => t,
-            None => return Vec::new(),
-        };
-        let bytes = source_text.as_bytes();
-        let mut result = Vec::new();
-        for c in &self.comment_ranges {
-            if c.pos >= from_pos && c.end <= to_pos {
-                // Check if standalone: only whitespace before it on the line
-                let mut line_start = c.pos as usize;
-                while line_start > 0
-                    && bytes[line_start - 1] != b'\n'
-                    && bytes[line_start - 1] != b'\r'
-                {
-                    line_start -= 1;
-                }
-                let before = &source_text[line_start..c.pos as usize];
-                if before.trim().is_empty() {
-                    let text = c.get_text(source_text);
-                    if !text.is_empty() {
-                        result.push(IRNode::Raw(text.to_string().into()));
-                    }
-                }
-            }
-            if c.pos >= to_pos {
-                break;
-            }
-        }
-        result
-    }
-
-    /// Extract a trailing comment within a statement's span.
-    ///
-    /// In our parser, `node.end` includes trailing trivia, so comments appear
-    /// WITHIN `[stmt_pos, stmt_end)` rather than after `stmt_end`. This method
-    /// finds comments within the span that have code on the same line before them
-    /// (i.e., they're trailing comments, not standalone leading comments).
-    fn extract_trailing_comment_in_stmt(&self, stmt_pos: u32, stmt_end: u32) -> Option<String> {
-        let source_text = self.source_text?;
-        let bytes = source_text.as_bytes();
-
-        for c in &self.comment_ranges {
-            if c.pos >= stmt_pos && c.end <= stmt_end {
-                // Check if there's non-whitespace code before this comment on the same line
-                let mut line_start = c.pos as usize;
-                while line_start > 0
-                    && bytes[line_start - 1] != b'\n'
-                    && bytes[line_start - 1] != b'\r'
-                {
-                    line_start -= 1;
-                }
-                let before_comment = &source_text[line_start..c.pos as usize];
-                if !before_comment.trim().is_empty() {
-                    let text = c.get_text(source_text);
-                    if !text.is_empty() {
-                        return Some(text.to_string());
-                    }
-                }
-            }
-            if c.pos >= stmt_end {
-                break;
-            }
-        }
-        None
-    }
-
-    fn extract_namespace_trailing_comment(&self, body_idx: NodeIndex) -> Option<String> {
-        let source_text = self.source_text?;
-        let body_node = self.arena.get(body_idx)?;
-        let pos = self.find_module_block_close_pos(body_node)? as usize;
-
-        let comments = crate::emitter::get_trailing_comment_ranges(source_text, pos + 1);
-        if comments.is_empty() {
-            return None;
-        }
-
-        Some(
-            comments
-                .iter()
-                .map(|comment| source_text[comment.pos as usize..comment.end as usize].to_string())
-                .collect::<Vec<_>>()
-                .join(" "),
-        )
-    }
-
-    fn find_module_block_close_pos(&self, body_node: &Node) -> Option<u32> {
-        let source_text = self.source_text?;
-        let bytes = source_text.as_bytes();
-        let limit = std::cmp::min(body_node.end as usize, bytes.len());
-        let mut pos = body_node.pos as usize;
-        while pos < limit && bytes.get(pos) != Some(&b'{') {
-            pos += 1;
-        }
-        if pos >= limit {
-            return None;
-        }
-
-        let mut depth = 0u32;
-        while pos < limit {
-            match bytes[pos] {
-                b'{' => {
-                    depth += 1;
-                    pos += 1;
-                }
-                b'}' => {
-                    depth = depth.saturating_sub(1);
-                    if depth == 0 {
-                        return Some(pos as u32);
-                    }
-                    pos += 1;
-                }
-                b'/' if pos + 1 < limit && bytes[pos + 1] == b'/' => {
-                    pos += 2;
-                    while pos < limit && !matches!(bytes[pos], b'\n' | b'\r') {
-                        pos += 1;
-                    }
-                }
-                b'/' if pos + 1 < limit && bytes[pos + 1] == b'*' => {
-                    pos += 2;
-                    while pos + 1 < limit && !(bytes[pos] == b'*' && bytes[pos + 1] == b'/') {
-                        pos += 1;
-                    }
-                    pos = std::cmp::min(pos + 2, limit);
-                }
-                b'\'' | b'"' | b'`' => {
-                    let quote = bytes[pos];
-                    pos += 1;
-                    while pos < limit {
-                        if bytes[pos] == b'\\' {
-                            pos = std::cmp::min(pos + 2, limit);
-                        } else if bytes[pos] == quote {
-                            pos += 1;
-                            break;
-                        } else {
-                            pos += 1;
-                        }
-                    }
-                }
-                _ => pos += 1,
-            }
-        }
-        None
-    }
-
-    fn trailing_same_line_comment_end_after(&self, pos: u32) -> Option<u32> {
-        let source_text = self.source_text?;
-        crate::emitter::get_trailing_comment_ranges(source_text, pos as usize)
-            .last()
-            .map(|comment| comment.end)
     }
 
     /// Transform a namespace declaration to IR
@@ -1752,400 +1502,6 @@ impl<'a> NamespaceES5Transformer<'a> {
 
         groups.reverse();
         groups.into_iter().flatten().collect()
-    }
-
-    fn rewrite_const_enum_accesses(&self, nodes: &mut [IRNode], namespace_path: &[String]) {
-        if self.const_enum_values.is_empty() {
-            return;
-        }
-
-        for node in nodes {
-            self.rewrite_const_enum_accesses_in_node(node, namespace_path);
-        }
-    }
-
-    fn rewrite_const_enum_accesses_in_node(&self, node: &mut IRNode, namespace_path: &[String]) {
-        if let Some(replacement) = self.const_enum_replacement(node, namespace_path) {
-            *node = replacement;
-            return;
-        }
-
-        match node {
-            IRNode::BinaryExpr { left, right, .. }
-            | IRNode::LogicalOr { left, right }
-            | IRNode::LogicalAnd { left, right } => {
-                self.rewrite_const_enum_accesses_in_node(left, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(right, namespace_path);
-            }
-            IRNode::PrefixUnaryExpr { operand, .. }
-            | IRNode::PostfixUnaryExpr { operand, .. }
-            | IRNode::Parenthesized(operand)
-            | IRNode::SpreadElement(operand)
-            | IRNode::ExpressionStatement(operand)
-            | IRNode::ThrowStatement(operand)
-            | IRNode::PrivateFieldGet {
-                receiver: operand, ..
-            }
-            | IRNode::PrivateStaticFieldGet {
-                receiver: operand, ..
-            }
-            | IRNode::PrivateFieldIn { obj: operand, .. } => {
-                self.rewrite_const_enum_accesses_in_node(operand, namespace_path);
-            }
-            IRNode::CallExpr { callee, arguments }
-            | IRNode::NewExpr {
-                callee, arguments, ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(callee, namespace_path);
-                for arg in arguments {
-                    self.rewrite_const_enum_accesses_in_node(arg, namespace_path);
-                }
-            }
-            IRNode::PropertyAccess { object, .. } => {
-                self.rewrite_const_enum_accesses_in_node(object, namespace_path);
-            }
-            IRNode::ElementAccess { object, index } => {
-                self.rewrite_const_enum_accesses_in_node(object, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(index, namespace_path);
-            }
-            IRNode::ConditionalExpr {
-                condition,
-                when_true,
-                when_false,
-            } => {
-                self.rewrite_const_enum_accesses_in_node(condition, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(when_true, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(when_false, namespace_path);
-            }
-            IRNode::CommaExpr(items)
-            | IRNode::CommaExprMultiline(items)
-            | IRNode::CommaExprMultilineFlat(items)
-            | IRNode::ArrayLiteral(items)
-            | IRNode::VarDeclList(items)
-            | IRNode::Block(items)
-            | IRNode::Sequence(items)
-            | IRNode::StaticBlockIIFE { statements: items } => {
-                for item in items {
-                    self.rewrite_const_enum_accesses_in_node(item, namespace_path);
-                }
-            }
-            IRNode::ObjectLiteral { properties, .. } => {
-                for property in properties {
-                    if let IRPropertyKey::Computed(key) = &mut property.key {
-                        self.rewrite_const_enum_accesses_in_node(key, namespace_path);
-                    }
-                    self.rewrite_const_enum_accesses_in_node(&mut property.value, namespace_path);
-                }
-            }
-            IRNode::FunctionExpr {
-                parameters, body, ..
-            }
-            | IRNode::FunctionDecl {
-                parameters, body, ..
-            } => {
-                for param in parameters {
-                    if let Some(default_value) = &mut param.default_value {
-                        self.rewrite_const_enum_accesses_in_node(default_value, namespace_path);
-                    }
-                }
-                for item in body {
-                    self.rewrite_const_enum_accesses_in_node(item, namespace_path);
-                }
-            }
-            IRNode::VarDecl { initializer, .. } => {
-                if let Some(initializer) = initializer {
-                    self.rewrite_const_enum_accesses_in_node(initializer, namespace_path);
-                }
-            }
-            IRNode::ReturnStatement(expr) => {
-                if let Some(expr) = expr {
-                    self.rewrite_const_enum_accesses_in_node(expr, namespace_path);
-                }
-            }
-            IRNode::IfStatement {
-                condition,
-                then_branch,
-                else_branch,
-            } => {
-                self.rewrite_const_enum_accesses_in_node(condition, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(then_branch, namespace_path);
-                if let Some(else_branch) = else_branch {
-                    self.rewrite_const_enum_accesses_in_node(else_branch, namespace_path);
-                }
-            }
-            IRNode::SwitchStatement { expression, cases } => {
-                self.rewrite_const_enum_accesses_in_node(expression, namespace_path);
-                for case in cases {
-                    if let Some(test) = &mut case.test {
-                        self.rewrite_const_enum_accesses_in_node(test, namespace_path);
-                    }
-                    for statement in &mut case.statements {
-                        self.rewrite_const_enum_accesses_in_node(statement, namespace_path);
-                    }
-                }
-            }
-            IRNode::ForStatement {
-                initializer,
-                condition,
-                incrementor,
-                body,
-            } => {
-                if let Some(initializer) = initializer {
-                    self.rewrite_const_enum_accesses_in_node(initializer, namespace_path);
-                }
-                if let Some(condition) = condition {
-                    self.rewrite_const_enum_accesses_in_node(condition, namespace_path);
-                }
-                if let Some(incrementor) = incrementor {
-                    self.rewrite_const_enum_accesses_in_node(incrementor, namespace_path);
-                }
-                self.rewrite_const_enum_accesses_in_node(body, namespace_path);
-            }
-            IRNode::ForInOfStatement {
-                initializer,
-                expression,
-                body,
-                ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(initializer, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(expression, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(body, namespace_path);
-            }
-            IRNode::WhileStatement { condition, body }
-            | IRNode::DoWhileStatement { body, condition } => {
-                self.rewrite_const_enum_accesses_in_node(condition, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(body, namespace_path);
-            }
-            IRNode::TryStatement {
-                try_block,
-                catch_clause,
-                finally_block,
-            } => {
-                self.rewrite_const_enum_accesses_in_node(try_block, namespace_path);
-                if let Some(catch_clause) = catch_clause {
-                    for statement in &mut catch_clause.body {
-                        self.rewrite_const_enum_accesses_in_node(statement, namespace_path);
-                    }
-                }
-                if let Some(finally_block) = finally_block {
-                    self.rewrite_const_enum_accesses_in_node(finally_block, namespace_path);
-                }
-            }
-            IRNode::LabeledStatement { statement, .. } => {
-                self.rewrite_const_enum_accesses_in_node(statement, namespace_path);
-            }
-            IRNode::ES5ClassIIFE {
-                base_class,
-                body,
-                computed_prop_temp_inits,
-                deferred_static_blocks,
-                ..
-            }
-            | IRNode::ES5ClassAssignment {
-                base_class,
-                body,
-                computed_prop_temp_inits,
-                deferred_static_blocks,
-                ..
-            } => {
-                if let Some(base_class) = base_class {
-                    self.rewrite_const_enum_accesses_in_node(base_class, namespace_path);
-                }
-                for item in body {
-                    self.rewrite_const_enum_accesses_in_node(item, namespace_path);
-                }
-                for item in computed_prop_temp_inits {
-                    self.rewrite_const_enum_accesses_in_node(item, namespace_path);
-                }
-                for item in deferred_static_blocks {
-                    self.rewrite_const_enum_accesses_in_node(item, namespace_path);
-                }
-            }
-            IRNode::ES5ClassApply {
-                factory,
-                base_class,
-            } => {
-                self.rewrite_const_enum_accesses_in_node(factory, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(base_class, namespace_path);
-            }
-            IRNode::PrototypeMethod {
-                method_name,
-                function,
-                ..
-            }
-            | IRNode::StaticMethod {
-                method_name,
-                function,
-                ..
-            } => {
-                if let crate::transforms::ir::IRMethodName::Computed(name) = method_name {
-                    self.rewrite_const_enum_accesses_in_node(name, namespace_path);
-                }
-                self.rewrite_const_enum_accesses_in_node(function, namespace_path);
-            }
-            IRNode::DefineProperty {
-                target,
-                property_name,
-                descriptor,
-                ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(target, namespace_path);
-                if let crate::transforms::ir::IRMethodName::Computed(name) = property_name {
-                    self.rewrite_const_enum_accesses_in_node(name, namespace_path);
-                }
-                if let Some(get) = &mut descriptor.get {
-                    self.rewrite_const_enum_accesses_in_node(get, namespace_path);
-                }
-                if let Some(set) = &mut descriptor.set {
-                    self.rewrite_const_enum_accesses_in_node(set, namespace_path);
-                }
-                if let Some(value) = &mut descriptor.value {
-                    self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-                }
-            }
-            IRNode::AwaiterCall {
-                this_arg,
-                generator_body,
-                ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(this_arg, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(generator_body, namespace_path);
-            }
-            IRNode::GeneratorBody { cases, .. } => {
-                for case in cases {
-                    for statement in &mut case.statements {
-                        self.rewrite_const_enum_accesses_in_node(statement, namespace_path);
-                    }
-                }
-            }
-            IRNode::GeneratorOp { value, .. } => {
-                if let Some(value) = value {
-                    self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-                }
-            }
-            IRNode::IfBreak { condition, .. } => {
-                self.rewrite_const_enum_accesses_in_node(condition, namespace_path);
-            }
-            IRNode::PrivateFieldSet {
-                receiver, value, ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(receiver, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-            }
-            IRNode::PrivateStaticFieldSet {
-                receiver,
-                state,
-                value,
-                ..
-            } => {
-                self.rewrite_const_enum_accesses_in_node(receiver, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(state, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-            }
-            IRNode::WeakMapSet { key, value, .. } => {
-                self.rewrite_const_enum_accesses_in_node(key, namespace_path);
-                self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-            }
-            IRNode::EnumIIFE { members, .. } => {
-                for member in members {
-                    if let EnumMemberValue::Computed(expr) = &mut member.value {
-                        self.rewrite_const_enum_accesses_in_node(expr, namespace_path);
-                    }
-                }
-            }
-            IRNode::NamespaceIIFE {
-                body, name_parts, ..
-            } => {
-                let nested_namespace_path = name_parts
-                    .iter()
-                    .map(|part| part.as_ref().to_string())
-                    .collect::<Vec<_>>();
-                for item in body {
-                    self.rewrite_const_enum_accesses_in_node(item, &nested_namespace_path);
-                }
-            }
-            IRNode::NamespaceExport { value, .. } => {
-                self.rewrite_const_enum_accesses_in_node(value, namespace_path);
-            }
-            _ => {}
-        }
-    }
-
-    fn const_enum_replacement(&self, node: &IRNode, namespace_path: &[String]) -> Option<IRNode> {
-        let IRNode::PropertyAccess { object, property } = node else {
-            return None;
-        };
-        let enum_path = ir_access_path(object)?;
-        let values = self.lookup_const_enum_values_for_ir(&enum_path, namespace_path)?;
-        let value = values.get(property.as_ref())?;
-        let literal = value.to_js_literal();
-        if self.remove_comments {
-            return Some(IRNode::Raw(literal.into()));
-        }
-        Some(IRNode::Raw(
-            format!("{literal} /* {enum_path}.{property} */").into(),
-        ))
-    }
-
-    fn lookup_const_enum_values_for_ir(
-        &self,
-        enum_path: &str,
-        namespace_path: &[String],
-    ) -> Option<&FxHashMap<String, EnumValue>> {
-        if let Some(values) = self.lookup_const_enum_values_direct(enum_path) {
-            return Some(values);
-        }
-        let namespace_prefix = namespace_path.join(".");
-        if !namespace_prefix.is_empty() {
-            let scoped_path = format!("{namespace_prefix}.{enum_path}");
-            if let Some(values) = self.lookup_const_enum_values_direct(&scoped_path) {
-                return Some(values);
-            }
-        }
-
-        if let Some(dot_pos) = enum_path.find('.') {
-            let first = &enum_path[..dot_pos];
-            let rest = &enum_path[dot_pos + 1..];
-            if !namespace_prefix.is_empty() {
-                let scoped_alias = format!("{namespace_prefix}.{first}");
-                if let Some(target) = self.const_enum_import_aliases.get(&scoped_alias) {
-                    let resolved = format!("{target}.{rest}");
-                    if let Some(values) = self.lookup_const_enum_values_direct(&resolved) {
-                        return Some(values);
-                    }
-                }
-            }
-            if let Some(target) = self.const_enum_import_aliases.get(first) {
-                let resolved = format!("{target}.{rest}");
-                return self.lookup_const_enum_values_direct(&resolved);
-            }
-        } else {
-            if !namespace_prefix.is_empty() {
-                let scoped_alias = format!("{namespace_prefix}.{enum_path}");
-                if let Some(target) = self.const_enum_import_aliases.get(&scoped_alias)
-                    && let Some(values) = self.lookup_const_enum_values_direct(target)
-                {
-                    return Some(values);
-                }
-            }
-            if let Some(target) = self.const_enum_import_aliases.get(enum_path) {
-                return self.lookup_const_enum_values_direct(target);
-            }
-        }
-
-        None
-    }
-
-    fn lookup_const_enum_values_direct(
-        &self,
-        enum_path: &str,
-    ) -> Option<&FxHashMap<String, EnumValue>> {
-        let entries = self.const_enum_values.get(enum_path)?;
-        entries
-            .iter()
-            .find(|entry| entry.scope_start == 0 && entry.scope_end == u32::MAX)
-            .or_else(|| entries.first())
-            .map(|entry| &entry.values)
     }
 
     /// Transform a function in namespace. When `force_export` is true, the function
